Farmers-Dashboard-Backend/requirements.txt
appdirs==1.4.4
autopep8==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
lxml==4.9.1
multitasking==0.0.11
numpy==1.23.4
pandas==1.5.1
psycopg2-binary==2.9.4
pycodestyle==2.9.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
requests==2.28.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.12
yfinance==0.1.85
sqlalchemy==2.0.20
python-multipart==0.0.6
uvicorn==0.23.2
fastapi==0.103.0
httpx==0.24.1
supervisor==4.2.5
jinja2==3.1.2
sqlalchemy==2.0.20
